Locate Odin DLLs in a standard way through [BEGIN]LIBPATH by default.
With this changeset, it's no more necessary to have the KLIB/Odin key
in the user registry. If, for some reason, a different version of Odin needs
to be loaded with the plugin, one may use the standard
BEGINLIBPATH/LIBPATHSTRICT=T trick for it.

Alternatively, the NPFLOS2_ODINDLLPATH environment variable may be
set to override the default DLL search location but this is essentially equivalent
to using the above two standard variables and kept there just for very lazy
people.