Deduce Win32 plugin location from wrapper DLL location.
By default, the wrapper DLL will try to load a Win32 DLL named 'npswf32.dll'
from the same directory where the wrapper DLL resides -- this is the most
common case. No registry entry (Flash10_Plugin\AdobePluginPath) is necessary
any longer and support for it is removed.

For the purposes of customization (testing, debugging), the Win32 DLL path may
be overridden with the NPFLOS2_WIN32DLL environment variable containing
the full path to the DLL including the file name.